Extract common component - google maps search
