+++ frontend/src/components/travelblock/TravelActivityBlockEdit.js
@@ -11,13 +11,7 @@ import {
   MuiPickersUtilsProvider,
 } from '@material-ui/pickers';
 
-import Paper from '@material-ui/core/Paper';
-import InputBase from '@material-ui/core/InputBase';
-import Divider from '@material-ui/core/Divider';
-import IconButton from '@material-ui/core/IconButton';
-import MenuIcon from '@material-ui/icons/Menu';
-import SearchIcon from '@material-ui/icons/Search';
-import DirectionsIcon from '@material-ui/icons/Directions';
+import GoogleMapSearch from '../common/GoogleMapSearch';
 
 const useCardStyles = makeStyles({
   card: {
@@ -34,29 +28,8 @@ const useCardStyles = makeStyles({
   },
 });
 
-const useMapStyles = makeStyles((theme) => ({
-  root: {
-    padding: '2px 4px',
-    display: 'flex',
-    alignItems: 'center',
-    width: 360,
-  },
-  input: {
-    marginLeft: theme.spacing(1),
-    flex: 1,
-  },
-  iconButton: {
-    padding: 10,
-  },
-  divider: {
-    height: 20,
-    margin: 4,
-  },
-}));
-
 export default function TravelActivityBlockEdit(props) {
   const cardClasses = useCardStyles();
-  const mapClasses = useMapStyles();
 
   const [selectedDate0, setSelectedDate0] = React.useState(new Date('2020-01-01T09:00:00'));
   const [selectedDate1, setSelectedDate1] = React.useState(new Date('2020-01-01T09:00:00'));
@@ -85,28 +58,7 @@ export default function TravelActivityBlockEdit(props) {
         <MuiPickersUtilsProvider utils={DateFnsUtils}>
           <Grid container direction="column" justify="space-around">
             <Grid container direction="row" justify="space-around" alignItems="center">
-              <Paper className={mapClasses.root}>
-                <IconButton className={mapClasses.iconButton} aria-label="menu">
-                  <MenuIcon />
-                </IconButton>
-                <InputBase
-                  className={mapClasses.input}
-                  placeholder="Search Google Maps"
-                  inputProps={{ 'aria-label': 'search google maps' }}
-                />
-                <IconButton
-                  className={mapClasses.iconButton}
-                  id="search"
-                  aria-label="search"
-                  onClick={() => props.searchHandler()}
-                >
-                  <SearchIcon />
-                </IconButton>
-                <Divider className={mapClasses.divider} orientation="vertical" />
-                <IconButton color="primary" className={mapClasses.iconButton} aria-label="directions">
-                  <DirectionsIcon />
-                </IconButton>
-              </Paper>
+              <GoogleMapSearch searchHandler={props.searchHandler} />
             </Grid>
             <Grid container direction="row" justify="space-around" alignItems="center">
               {containerId.map((i) => {
