+++ frontend/src/components/travelblock/TravelActivityBlockEdit.test.js
@@ -12,21 +12,5 @@ describe('<TravelActivityBlockEdit />', () => {
     const component = shallow(<TravelActivityBlockEdit />);
     const wrapperCard = component.find(Card);
     expect(wrapperCard.length).toBe(1);
-    const wrapperPaper = component.find(Paper);
-    expect(wrapperPaper.length).toBe(1);
   });
-  it('should request search', () => {
-    const searchHandler = jest.fn();
-    const component = shallow(<TravelActivityBlockEdit searchHandler={searchHandler} />);
-    const button = component.find('#search');
-    expect(button.length).toBe(1);
-    button.simulate('click');
-    expect(searchHandler).toHaveBeenCalledTimes(1);
-  });
-  // it('should render title', () => {
-  //   const title = 'Activity';
-  //   const component = shallow(<TravelActivityBlockEdit title={title} />);
-  //   const header = component.find(CardHeader);
-  //   expect(header.props().subheader).toBe(title);
-  // });
 });
